Some null checks & more Dir > Directory

+++ src/GitJanitor.IO/GitRepositoryHandler.cs
@@ -27,49 +27,52 @@ public class GitRepositoryHandler : IGitRepositoryHandler
         await Task.WhenAll(tasks);
     }
 
-    public async Task ArchiveAsync(string workingDir, string targetDirectory, string zipFileName)
+    public async Task ArchiveAsync(string workingDirectory, string? targetDirectory, string zipFileName)
     {
         await Task.Run(() =>
         {
-            // Create a .zip at the source directory
-            var zipPath = Path.Combine(workingDir, zipFileName);
-            File.Delete(zipPath);
-            ZipFile.CreateFromDirectory(workingDir, zipPath);
+            if (!string.IsNullOrEmpty(targetDirectory))
+            {
+                // Create a .zip at the source directory
+                var zipPath = Path.Combine(workingDirectory, zipFileName);
+                File.Delete(zipPath);
+                ZipFile.CreateFromDirectory(workingDirectory, zipPath);
 
-            // Move the .zip file to the target directory
-            var movedZipPath = Path.Combine(targetDirectory, zipFileName);
-            File.Delete(movedZipPath);
-            File.Move(zipPath, movedZipPath);
+                // Move the .zip file to the target directory
+                var movedZipPath = Path.Combine(targetDirectory, zipFileName);
+                File.Delete(movedZipPath);
+                File.Move(zipPath, movedZipPath);
 
-            // Deletes the current (working) directory
-            Directory.Delete(workingDir, true);
+                // Deletes the current (working) directory
+                Directory.Delete(workingDirectory, true);
+            }
         });
     }
 
-    public async Task DeleteAsync(string workingDir)
+    public async Task DeleteAsync(string workingDirectory)
     {
         await Task.Run(() =>
             { 
-                Directory.Delete(workingDir, true);
+                Directory.Delete(workingDirectory, true);
             }
         );
     }
 
-    public async Task MoveAsync(string workingDir, string targetDir, string dirName)
+    public async Task MoveAsync(string workingDirectory, string? targetDirectory, string directoryName)
     {
         await Task.Run(() =>
         {
-            if (Directory.Exists(targetDir))
+            if (Directory.Exists(targetDirectory) && !string.IsNullOrEmpty(targetDirectory))
             {
-                var targetDirPath = Path.Combine(targetDir, dirName);
-                if (Directory.Exists(targetDirPath))
-                    Console.WriteLine("Target directory already exists: " + targetDirPath);
+                var targetPath = Path.Combine(targetDirectory, directoryName);
+                if (Directory.Exists(targetPath))
+                    Console.WriteLine("Target directory already exists: " + targetPath);
                 else
-                    Directory.Move(workingDir, targetDirPath);
+                    Directory.Move(workingDirectory, targetPath);
             }
             else
             {
-                throw new NotFoundException($"Location: {workingDir} Does not exist.");
+                throw new NotFoundException($"Location: {workingDirectory} Does not exist.");
             }
         });
     }
